+++ lib/src/dart/extensions.dart
@@ -1,3 +1,7 @@
 export 'list_extension.dart';
 
 export 'map_extension.dart';
+
+export 'num_extension.dart';
+
+export 'int_extension.dart';