+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 1.2.0
 
+### Added
+- Added `suffix` num extensions
+- Added `getMonthName` & `getWeekDay` int extensions
 ### Fixed
 - Fix Text color extensions.
+- Fix `isUpper` & `isLower` String extensions
+- Fix `hideIf` & `showIf` Widget extensions
 
 ## 1.1.0